Add support for PPL queries in backend handler
